Effettua la conversione come double, non usare i float per poi convertire il risultato in double.

+++ energy_device.cpp
@@ -860,7 +860,7 @@ QDate EnergyDevice::getDateFromFrame(OpenMsg &msg)
 
 double EnergyConversions::convertToRawData(qint64 bt_bus_data, EnergyConversions::EnergyTypology type)
 {
-	float factor;
+	double factor;
 	switch (type)
 	{
 	case DEFAULT_ENERGY:
